Actually get Punditize working in example app
Turns out that the example app loads the `pundit` gem only in the test
environment. As a result, test behaves differently from development,
causing confusion. The main difference is that the `Punditize` module
loads only when the gem is available.

Here I'm fixing the situation by making `pundit` available across
all environments. However this is not a perfect solution...

The app's `ApplicationController` has the following method:

```
  def show_action?(action, resource)
    action != :index || resource != ::Order
  end
```

This is overridden by Punditize... but we only use Punditize in
`ProductsController`. So the "Orders" navigation link only appears
when we are visiting the "Products" section.

I'm going to follow this up with changes to how authorization works
in the example app, to remove this issue and bring some order to
the app in general.