Add support for exact search (#35)
We want to support search functionality in the emails admin dashboard in
web, but the underlying administrate search functionality always uses
substring matching and doesn't offer the option to use an equality
comparison. We'd like to use this in order to have much more efficient
and accurate queries, since those search terms are going to be
pre-populated in the app through clickable links in the app wherever we
use the `Email` field.

This change adds that functionality.

Blocks this PR -> https://github.com/rinsed-org/web/pull/14327